update the rest of redis code and get ready for packaging
